Rename `overlay::Content` trait to `Overlay`
The `Overlay` struct is now `overlay::Element`.
